Rename project main/script folder to main/app
