+++ src/test/output/statistics/results-links.csv
@@ -66,3 +66,10 @@
 =HYPERLINK(/tree/master/output-deletions;deletions),=HYPERLINK(/tree/master/output-deletions/files/deletions/10fe406824dd71ab9e14a3418d7bda4ab53e32a1;deletions/files/deletions/10fe406824dd71ab9e14a3418d7bda4ab53e32a1),false,0,false,0,1.0,2.0,1.0,2.449489742783178,0.0,0
 =HYPERLINK(/tree/master/output-sameline;sameline),=HYPERLINK(/tree/master/output-sameline/files/sameline/ef1a4c9095c04af5eab2e0c54f590c053c48d410;sameline/files/sameline/ef1a4c9095c04af5eab2e0c54f590c053c48d410),false,0,false,0,1.0,2.0,1.0,0.0,0.0,0
 =HYPERLINK(/tree/master/output-mergeconflict;mergeconflict),=HYPERLINK(/tree/master/output-mergeconflict/files/mergeconflict/c00294fb7135e5bdcd6f7c3417405aea75b66444;mergeconflict/files/mergeconflict/c00294fb7135e5bdcd6f7c3417405aea75b66444),false,2,true,1,1.0,10.0,1.0,0.0,0.0,0
+=HYPERLINK(/tree/master/output-deletions;deletions),=HYPERLINK(/tree/master/output-deletions/files/deletions/de96a282565510196a5b0d81fc92e91c1e909428;deletions/files/deletions/de96a282565510196a5b0d81fc92e91c1e909428),false,0,false,0,1.0,10.488088481701515,1.0,1.4142135623730951,0.0,0
+=HYPERLINK(/tree/master/output-deletions;deletions),=HYPERLINK(/tree/master/output-deletions/files/deletions/1598887a045a86222b6d96dd1357b6a08f232ffa;deletions/files/deletions/1598887a045a86222b6d96dd1357b6a08f232ffa),false,1,true,1,1.0,6.0,1.0,3.4641016151377544,0.0,0
+=HYPERLINK(/tree/master/output-deletions;deletions),=HYPERLINK(/tree/master/output-deletions/files/deletions/10fe406824dd71ab9e14a3418d7bda4ab53e32a1;deletions/files/deletions/10fe406824dd71ab9e14a3418d7bda4ab53e32a1),false,0,false,0,1.0,2.0,1.0,2.449489742783178,0.0,0
+=HYPERLINK(/tree/master/output-sameline;sameline),=HYPERLINK(/tree/master/output-sameline/files/sameline/ef1a4c9095c04af5eab2e0c54f590c053c48d410;sameline/files/sameline/ef1a4c9095c04af5eab2e0c54f590c053c48d410),false,0,false,0,1.0,2.0,1.0,0.0,0.0,0
+=HYPERLINK(/tree/master/output-mergeconflict;mergeconflict),=HYPERLINK(/tree/master/output-mergeconflict/files/mergeconflict/c00294fb7135e5bdcd6f7c3417405aea75b66444;mergeconflict/files/mergeconflict/c00294fb7135e5bdcd6f7c3417405aea75b66444),false,2,true,1,1.0,10.0,1.0,0.0,0.0,0
+=HYPERLINK(/tree/master/output-mergeconflict;mergeconflict),=HYPERLINK(/tree/master/output-mergeconflict/files/mergeconflict/ea29fc17bdf20b40208ae39f88fd8aa2b2a5322b;mergeconflict/files/mergeconflict/ea29fc17bdf20b40208ae39f88fd8aa2b2a5322b),false,2,true,2,1.0,6.0,2.0,0.0,0.0,0
+=HYPERLINK(/tree/master/output-mergeconflict;mergeconflict),=HYPERLINK(/tree/master/output-mergeconflict/files/mergeconflict/b610225c36e1eb020b9899be907e06b96a19c1c0;mergeconflict/files/mergeconflict/b610225c36e1eb020b9899be907e06b96a19c1c0),false,1,true,1,1.0,2.0,1.0,0.0,0.0,0
